+++ gen/functions.cpp
@@ -22,6 +22,7 @@
 #include "gen/arrays.h"
 #include "gen/classes.h"
 #include "gen/dvalue.h"
+#include "gen/funcgenstate.h"
 #include "gen/function-inlining.h"
 #include "gen/inlineir.h"
 #include "gen/irstate.h"
@@ -863,24 +864,21 @@ void DtoDefineFunction(FuncDeclaration *fd, bool linkageAvailableExternally) {
   }
 
   IrFunction *irFunc = getIrFunc(fd);
-  IrFuncTy &irFty = irFunc->irFty;
 
   // debug info
   irFunc->diSubprogram = gIR->DBuilder.EmitSubProgram(fd);
 
-  Type *t = fd->type->toBasetype();
-  TypeFunction *f = static_cast<TypeFunction *>(t);
-  // assert(f->ctype);
-
-  llvm::Function *func = irFunc->func;
-
-  // is there a body?
-  if (fd->fbody == nullptr) {
+  if (!fd->fbody) {
     return;
   }
 
   IF_LOG Logger::println("Doing function body for: %s", fd->toChars());
-  gIR->functions.push_back(irFunc);
+  gIR->funcGenStates.emplace_back(new FuncGenState(*irFunc, *gIR));
+  auto &funcGen = gIR->funcGen();
+
+  const auto f = static_cast<TypeFunction *>(fd->type->toBasetype());
+  IrFuncTy &irFty = irFunc->irFty;
+  llvm::Function *func = irFunc->func;;
 
   const auto lwc = lowerFuncLinkage(fd);
   if (linkageAvailableExternally) {
@@ -917,7 +915,6 @@ void DtoDefineFunction(FuncDeclaration *fd, bool linkageAvailableExternally) {
   llvm::BasicBlock *beginbb =
       llvm::BasicBlock::Create(gIR->context(), "", func);
 
-  // assert(gIR->scopes.empty());
   gIR->scopes.push_back(IRScope(beginbb));
 
 // Set the FastMath options for this function scope.
@@ -932,7 +929,7 @@ void DtoDefineFunction(FuncDeclaration *fd, bool linkageAvailableExternally) {
   // matter at all
   llvm::Instruction *allocaPoint = new llvm::AllocaInst(
       LLType::getInt32Ty(gIR->context()), "alloca point", beginbb);
-  irFunc->allocapoint = allocaPoint;
+  funcGen.allocapoint = allocaPoint;
 
   // debug info - after all allocas, but before any llvm.dbg.declare etc
   gIR->DBuilder.EmitFuncStart(fd);
@@ -986,45 +983,38 @@ void DtoDefineFunction(FuncDeclaration *fd, bool linkageAvailableExternally) {
     defineParameters(irFty, *fd->parameters);
 
   // Initialize PGO state for this function
-  irFunc->pgo.assignRegionCounters(fd, irFunc->func);
+  funcGen.pgo.assignRegionCounters(fd, irFunc->func);
 
+  DtoCreateNestedContext(funcGen);
+
+  if (fd->vresult && !fd->vresult->nestedrefs.dim) // FIXME: not sure here :/
   {
-    ScopeStack scopeStack(gIR);
-    irFunc->scopes = &scopeStack;
-
-    DtoCreateNestedContext(fd);
-
-    if (fd->vresult && !fd->vresult->nestedrefs.dim) // FIXME: not sure here :/
-    {
-      DtoVarDeclaration(fd->vresult);
-    }
-
-    // D varargs: prepare _argptr and _arguments
-    if (f->linkage == LINKd && f->varargs == 1) {
-      // allocate _argptr (of type core.stdc.stdarg.va_list)
-      Type *const argptrType = Type::tvalist->semantic(fd->loc, fd->_scope);
-      LLValue *argptrMem = DtoAlloca(argptrType, "_argptr_mem");
-      irFunc->_argptr = argptrMem;
-
-      // initialize _argptr with a call to the va_start intrinsic
-      DLValue argptrVal(argptrType, argptrMem);
-      LLValue *llAp = gABI->prepareVaStart(&argptrVal);
-      llvm::CallInst::Create(GET_INTRINSIC_DECL(vastart), llAp, "",
-                             gIR->scopebb());
-
-      // copy _arguments to a memory location
-      irFunc->_arguments =
-          DtoAllocaDump(irFunc->_arguments, 0, "_arguments_mem");
-    }
-
-    irFunc->pgo.emitCounterIncrement(fd->fbody);
-    irFunc->pgo.setCurrentStmt(fd->fbody);
-
-    // output function body
-    Statement_toIR(fd->fbody, gIR);
-
-    irFunc->scopes = nullptr;
+    DtoVarDeclaration(fd->vresult);
   }
+
+  // D varargs: prepare _argptr and _arguments
+  if (f->linkage == LINKd && f->varargs == 1) {
+    // allocate _argptr (of type core.stdc.stdarg.va_list)
+    Type *const argptrType = Type::tvalist->semantic(fd->loc, fd->_scope);
+    LLValue *argptrMem = DtoAlloca(argptrType, "_argptr_mem");
+    irFunc->_argptr = argptrMem;
+
+    // initialize _argptr with a call to the va_start intrinsic
+    DLValue argptrVal(argptrType, argptrMem);
+    LLValue *llAp = gABI->prepareVaStart(&argptrVal);
+    llvm::CallInst::Create(GET_INTRINSIC_DECL(vastart), llAp, "",
+                           gIR->scopebb());
+
+    // copy _arguments to a memory location
+    irFunc->_arguments =
+        DtoAllocaDump(irFunc->_arguments, 0, "_arguments_mem");
+  }
+
+  funcGen.pgo.emitCounterIncrement(fd->fbody);
+  funcGen.pgo.setCurrentStmt(fd->fbody);
+
+  // output function body
+  Statement_toIR(fd->fbody, gIR);
 
   llvm::BasicBlock *bb = gIR->scopebb();
   if (pred_begin(bb) == pred_end(bb) &&
@@ -1057,14 +1047,15 @@ void DtoDefineFunction(FuncDeclaration *fd, bool linkageAvailableExternally) {
 
   // erase alloca point
   if (allocaPoint->getParent()) {
+    funcGen.allocapoint = nullptr;
     allocaPoint->eraseFromParent();
+    allocaPoint = nullptr;
   }
-  allocaPoint = nullptr;
-  gIR->func()->allocapoint = nullptr;
 
   gIR->scopes.pop_back();
 
-  gIR->functions.pop_back();
+  assert(&gIR->funcGen() == &funcGen);
+  gIR->funcGenStates.pop_back();
 }
 
 ////////////////////////////////////////////////////////////////////////////////
