+++ gen/irstate.h
@@ -15,16 +15,16 @@
 #ifndef LDC_GEN_IRSTATE_H
 #define LDC_GEN_IRSTATE_H
 
-#include "aggregate.h"
-#include "root.h"
-#include "ir/iraggr.h"
-#include "ir/irvar.h"
-#include "gen/dibuilder.h"
 #include <deque>
-#include <list>
+#include <memory>
 #include <set>
 #include <sstream>
 #include <vector>
+#include "aggregate.h"
+#include "root.h"
+#include "gen/dibuilder.h"
+#include "ir/iraggr.h"
+#include "ir/irvar.h"
 #include "llvm/ADT/StringMap.h"
 #include "llvm/ProfileData/InstrProfReader.h"
 #include "llvm/IR/CallSite.h"
@@ -35,7 +35,7 @@ class TargetMachine;
 class IndexedInstrProfReader;
 }
 
-// global ir state for current module
+class FuncGenState;
 struct IRState;
 struct TargetABI;
 
@@ -107,6 +107,7 @@ struct IRAsmBlock {
 // represents the module
 struct IRState {
   IRState(const char *name, llvm::LLVMContext &context);
+  ~IRState();
 
   llvm::Module module;
   llvm::LLVMContext &context() const { return module.getContext(); }
@@ -116,11 +117,12 @@ struct IRState {
   LLStructType *mutexType;
   LLStructType *moduleRefType;
 
-  // functions
-  typedef std::vector<IrFunction *> FunctionVector;
-  FunctionVector functions;
+  // Stack of currently codegen'd functions (more than one for lambdas or other
+  // nested functions, inlining-only codegen'ing, etc.), and some convenience
+  // accessors for the top-most one.
+  std::vector<std::unique_ptr<FuncGenState>> funcGenStates;
+  FuncGenState &funcGen();
   IrFunction *func();
-
   llvm::Function *topfunc();
   llvm::Instruction *topallocapoint();
 
@@ -197,5 +199,7 @@ struct IRState {
 };
 
 void Statement_toIR(Statement *s, IRState *irs);
+
+bool useMSVCEH();
 
 #endif // LDC_GEN_IRSTATE_H
